+++ model_navigator/torch/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2021-2023, NVIDIA CORPORATION. All rights reserved.
+# Copyright (c) 2021-2024, NVIDIA CORPORATION. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -46,6 +46,7 @@ from model_navigator.pipelines.builders import (
     verify_builder,
 )
 from model_navigator.pipelines.builders.find_device_max_batch_size import find_device_max_batch_size_builder
+from model_navigator.pipelines.builders.torch import torch_exportedprogram_builder
 from model_navigator.pipelines.wrappers.optimize import optimize_pipeline
 from model_navigator.runners.base import NavigatorRunner
 from model_navigator.runners.utils import default_runners, filter_runners
@@ -144,6 +145,7 @@ def optimize(
         preprocessing_builder,
         torch_export_builder,
         find_device_max_batch_size_builder,
+        torch_exportedprogram_builder,
         torch_conversion_builder,
         torch_tensorrt_conversion_builder,
         tensorrt_conversion_builder,